+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 10.7
 - Escape SanitizeHTML() & SanitizeMarkDown() before saving to DB, program wide
 - Always use UserStudentID() instead of `$_SESSION['STUDENT_ID']` in PortalPollsNotes.php
 - SQL use EXISTS(SELECT 1) instead of `EXISTS(SELECT *)` in EditReportCardGrades.php, Rollover.php & Moodle/functions.php
+- Remove trailing seconds :00 & add sorting HTML comment to datetime in Date.php
 
 Changes in 10.6.3
 -----------------